+++ Userland/SampleCodeModule/include/stdlib.h
@@ -1,5 +1,7 @@
 #ifndef _STDLIB_H
 #define _STDLIB_H
+#include <stdint.h>
+
 /**
  * @brief Convierte un arreglo de caracteres en numero entero en formato decimal
  * @param s Cadena de caracteres
@@ -30,5 +32,5 @@ int strtoh(char* s, char ** end);
  * @param base Base del numero que se desea convertir
  * @return Longitud del buffer (numero que se convirtio en cadena de caracteres)
  */
-int itoa(int n, char* buffer, int base);
+int itoa(uint64_t n, char* buffer, int base);
 #endif
